+++ src/IDBridge/Public/Google/Get-GoogleUserGroupsToUpdate.ps1
@@ -59,7 +59,7 @@ function Get-GoogleUserGroupsToUpdate {
         }
 
         if ($userGroupsAdd.Count -gt 0) {
-            Write-Log -Message "Google: Information that needs updating - Add Groups: $($item.personID) $($item.NameFirst) $($item.NameLast): $($userGroupsAdd -join ', ')"
+            Write-Log -Message "Google: Proposed: Add Groups: $($item.personID) $($item.NameFirst) $($item.NameLast): $($userGroupsAdd -join ', ')"
             $itemListAdd += [PSCustomObject]@{
                 PersonID = $item.PersonID
                 GoogleCurrentUserID = $item.GoogleCurrentUserID
@@ -77,7 +77,7 @@ function Get-GoogleUserGroupsToUpdate {
         }
 
         if ($userGroupsRemove.Count -gt 0) {
-            Write-Log -Message "Google: Information that needs updating - Remove Groups: $($item.personID) $($item.NameFirst) $($item.NameLast): $($userGroupsRemove -join ', ')"
+            Write-Log -Message "Google: Proposed: Remove Groups: $($item.personID) $($item.NameFirst) $($item.NameLast): $($userGroupsRemove -join ', ')"
             $itemListRemove += [PSCustomObject]@{
                 PersonID = $item.PersonID
                 GoogleCurrentUserID = $item.GoogleCurrentUserID