+++ src/IDBridge/Public/Google/Get-GoogleUsersToUpdate.ps1
@@ -103,7 +103,7 @@ function Get-GoogleUsersToUpdate {
         if ($itemUpdateSplat.Count -gt 0) {
             $itemUpdateSplat["GoogleUserID"] = $item.GoogleCurrentUserID
 
-            Write-Log -Message ("Google: Information that needs updating for: " + $item.UPN + " - " + $item.personID)
+            Write-Log -Message ("Google: Proposed: Update User: " + $item.UPN + " - " + $item.personID)
             Write-Log -Message ($itemUpdateSplat | ConvertTo-Json -Compress)
 
             $itemUpdateList += [PSCustomObject]@{